Extract AgentTab and MissionsTab from Dashboard
Split the monolithic Dashboard into focused components:
- AgentTab: agent identity card, cover story, cover story modal
- MissionsTab: phase-grouped mission list with completion status
- Dashboard: tab shell, session/mission data fetching, mission modal
